fix: prevent image deletion when image not found

+++ app/modules/database/controllers/ImagesController.php
@@ -212,34 +212,41 @@ class Database_ImagesController extends Pas_Controller_Action_Admin
             $id = (int)$this->_request->getPost('id');
             $del = $this->_request->getPost('del');
             $type  = $this->_request->getPost('recordtype');
+
             if ($del == 'Yes' && $id > 0) {
                 $imagedata = $this->_images->getFileName($id, $type);
                 $filename = $imagedata['0']['f'];
-                $splitf = explode('.', $filename);
-                $spf = $splitf['0'];
-                $imagedir = rtrim($imagedata['0']['imagedir'],'/') . '/';
-                $imagenumber = $imagedata['0']['imageID'];
-                $zoom = './' . $imagedir . 'zoom/' . $spf . '_zdata';
-                $thumb = rtrim(IMAGE_PATH, '/') . '/' . 'thumbnails/' . $imagenumber . '.jpg';
-                $small = './' . $imagedir . 'small/' . $filename;
-                $display = './' . $imagedir . 'display/' . $filename;
-                $medium = './' . $imagedir . 'medium/' . $filename;
-                $original = './' . $imagedir . $filename;
-                $where = 'imageID = ' . $id;
-                $this->_images->delete($where);
-                $this->_helper->solrUpdater->deleteById('images', $id);
-                $linked = new FindsImages();
-                $wherelinks = array();
-                $wherelinks[] = $linked->getAdapter()->quoteInto('image_id = ?', $imagedata['0']['secuid']);
-                $linked->delete($wherelinks);
-                $this->getFlash()->addMessage('Image and metadata deleted');
-                $images = array($thumb, $display, $small, $medium, $original, $zoom);
 
-                //Delete images from server
-                $this->unlinker($images);
-                $this->deleteImageOriginals($original);
+                if(!(empty($filename))) {
+                    $splitf = explode('.', $filename);
+                    $spf = $splitf['0'];
+                    $imagedir = rtrim($imagedata['0']['imagedir'],'/') . '/';
+                    $imagenumber = $imagedata['0']['imageID'];
+                    $zoom = './' . $imagedir . 'zoom/' . $spf . '_zdata';
+                    $thumb = rtrim(IMAGE_PATH, '/') . '/' . 'thumbnails/' . $imagenumber . '.jpg';
+                    $small = './' . $imagedir . 'small/' . $filename;
+                    $display = './' . $imagedir . 'display/' . $filename;
+                    $medium = './' . $imagedir . 'medium/' . $filename;
+                    $original = './' . $imagedir . $filename;
+                    $where = 'imageID = ' . $id;
+                    $this->_images->delete($where);
+                    $this->_helper->solrUpdater->deleteById('images', $id);
+                    $linked = new FindsImages();
+                    $wherelinks = array();
+                    $wherelinks[] = $linked->getAdapter()->quoteInto('image_id = ?', $imagedata['0']['secuid']);
+                    $linked->delete($wherelinks);
+                    $this->getFlash()->addMessage('Image and metadata deleted');
+                    $images = array($thumb, $display, $small, $medium, $original, $zoom);
 
-                $this->_helper->solrUpdater->update('objects', $imagedata['0']['id'], $recordtype);
+                    //Delete images from server
+                    $this->unlinker($images);
+                    $this->deleteImageOriginals($original);
+
+                    //Update SOLR with image deletion
+                    $this->_helper->solrUpdater->update('objects', $imagedata['0']['id'], $recordtype);
+                } else {
+                    $this->getFlash()->addMessage('Image not found. Please contact support.');
+                }
             }
             $this->redirect('/database/myscheme/myimages/');
         } else {
